feat : 구독권을 생성하는 기능 추가 및 테스트 작성
- data.sql로 데이터 미리 저장

+++ src/main/java/com/example/subscription/entity/SubscriptionProduct.java
@@ -16,13 +16,13 @@ public class SubscriptionProduct {
 
     private String name;
 
-    private long amount;
+    private Long amount;
 
-    private int period;
+    private Long period;
 
-    private int chatTimes;
+    private Integer chatTimes;
 
-    private int unitStudyTimes;
+    private Integer unitStudyTimes;
 
     @Enumerated(EnumType.STRING)
     private SubscriptionType subscriptionType;
